requirements.txt
torch==2.3.0 --index-url https://download.pytorch.org/whl/cu121
transformers==4.41.2
accelerate==0.30.1
bitsandbytes==0.43.1   # 8-bit loading
sentence-transformers==2.7.0
chromadb==0.5.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.4
pandas==2.2.2
scikit-learn==1.5.0
pydantic==2.7.4
python-dotenv==1.0.1